+++ functions/lib/db/index.js
@@ -1,13 +1,14 @@
 const pg = require("./pg/pg")
+const otp = require("./otp")
+
 const Accounts = require("./accounts")
-const Otp = require("./otp")
 const ErrorCodes = require("./error_codes")
 const Rooms = require("./rooms")
 const Profile = require("./profile")
 
 module.exports = {
   pg: pg,
-  Otp: Otp,
+  otp: otp,
   Accounts: Accounts,
   ErrorCodes: ErrorCodes,
   Rooms: Rooms,