+++ functions/lib/index.js
@@ -7,12 +7,13 @@
 
 require("./globals")
 
-global.util = require("./util/index.js")
-global.db = require("./db/index.js")
+global.lib = {}
 
-global.lib = {
-  util: util,
-  db: db
-}
+global.lib.util = require("./util/index.js")
+global.lib.db = require("./db/index.js")
+global.lib.email = require("./email/index.js")
+
+global.util = lib.util
+global.db = lib.db
 
 module.exports = lib