+++ functions/request_create_account.js
@@ -11,10 +11,6 @@ Sign up flow
 5. Create/store session
 */
 
-const sendOtpEmail = async (params, signupUrl) => {
-  // TODO:  send email
-  console.log(`Sending ${signupUrl} to email: ${params.email}`)
-}
 
 module.exports.handler = async (event, context, callback) => {
   // We only care about POSTs with body data
@@ -35,7 +31,7 @@ module.exports.handler = async (event, context, callback) => {
   const existingCreateRequest = await accounts.getLatestAccountCreateRequest(params.email)
 
   if(existingCreateRequest) {
-    if(!accounts.isExpired(existingCreateRequest)) {
+    if(!lib.db.otp.isExpired(existingCreateRequest)) {
       return lib.util.http.fail(callback, "Email already registered. Check your email for a verification link.")
     }
   }
@@ -50,8 +46,8 @@ module.exports.handler = async (event, context, callback) => {
 
   const signupUrl = accounts.getSignupUrl(lib.util.env.appUrl(event, context), createRequest)
 
-  await sendOtpEmail(params, signupUrl)
+  await lib.email.account.sendSignupOtpEmail(params.email, params.firstName, signupUrl)
 
   await accounts.cleanup()
-  return lib.util.http.succeed(callback, {signupUrl: signupUrl});
+  return lib.util.http.succeed(callback, {});
 }