+++ functions/resolve_init_session.js
@@ -1,7 +1,7 @@
 const lib = require("lib");
 lib.util.env.init(require("./env.json"))
 
-const handleAccountCreateFailure = (errorCode, callback) => {
+const handleLoginFailure = (errorCode, callback) => {
   switch(errorCode) {
     case lib.db.ErrorCodes.otp.INVALID_OTP:
       return util.http.fail(callback, "Invalid one-time passcode.");
@@ -9,7 +9,7 @@ const handleAccountCreateFailure = (errorCode, callback) => {
       return util.http.fail(callback, "Sorry, this link has expired. Please sign up again.");
     case lib.db.ErrorCodes.otp.RESOLVED_OTP:
       return util.http.fail(callback, "It seems this link has already been used to log in. Please try again.");
-    case lib.db.ErrorCodes.otp.UNEXPECTER_ERROR:
+    case lib.db.ErrorCodes.UNEXPECTER_ERROR:
       // TODO: ERROR_LOGGING
       return util.http.fail(callback, "An unexpected error happened. Please try again.");
     default:
@@ -35,7 +35,7 @@ module.exports.handler = async (event, context, callback) => {
 
   const result = await accounts.resolveLoginRequest(uid, otp)
   if(result.error != null) {
-    return handleAccountCreateFailure(result.error, callback)
+    return handleLoginFailure(result.error, callback)
   }
 
   const session = result.session