+++ src/components/VideoProvider/useLocalTracks/useLocalTracks.ts
@@ -5,12 +5,16 @@
  * - Updated to latest starter app code
  * - Add local data track
  * - Video off by default
+ * Aug 28, 2020 WQP
+ * - Add error handler/message for failing to get the initial audio track.
  */
 import { DEFAULT_VIDEO_CONSTRAINTS } from '../../../constants';
 import { useCallback, useEffect, useState } from 'react';
 import Video, { LocalVideoTrack, LocalAudioTrack, CreateLocalTrackOptions, LocalDataTrack } from 'twilio-video';
 
-export default function useLocalTracks() {
+import { ErrorCallback } from '../../../types';
+
+export default function useLocalTracks(onError: ErrorCallback) {
   const [audioTrack, setAudioTrack] = useState<LocalAudioTrack>();
   const [videoTrack, setVideoTrack] = useState<LocalVideoTrack>();
   const [dataTrack, setDataTrack] = useState<LocalDataTrack>();
@@ -23,7 +27,7 @@ export default function useLocalTracks() {
       options.deviceId = { exact: deviceId };
     }
 
-    return Video.createLocalAudioTrack(options).then(newTrack => {
+    return Video.createLocalAudioTrack(options).then((newTrack) => {
       setAudioTrack(newTrack);
       return newTrack;
     });
@@ -41,7 +45,7 @@ export default function useLocalTracks() {
       ...newOptions,
     };
 
-    return Video.createLocalVideoTrack(options).then(newTrack => {
+    return Video.createLocalVideoTrack(options).then((newTrack) => {
       setVideoTrack(newTrack);
       return newTrack;
     });
@@ -64,22 +68,35 @@ export default function useLocalTracks() {
       // },
       audio: true,
     })
-      .then(tracks => {
-        const videoTrack = tracks.find(track => track.kind === 'video');
-        const audioTrack = tracks.find(track => track.kind === 'audio');
+      .then((tracks) => {
+        const videoTrack = tracks.find((track) => track.kind === 'video');
+        const audioTrack = tracks.find((track) => track.kind === 'audio');
         if (videoTrack) {
           setVideoTrack(videoTrack as LocalVideoTrack);
         }
         if (audioTrack) {
           setAudioTrack(audioTrack as LocalAudioTrack);
         }
-
-        setDataTrack(new LocalDataTrack());
+      })
+      .catch((err) => {
+        const msg =
+          // Currently only attempting to connect audio by default, so let the messaging reflect that.
+          // May need to update the error string if we decide to enable video by default later on.
+          err.message === 'Permission denied' || 'Permission dismissed'
+            ? 'We were unable to connect to your microphone. Please grant microphone access to enable audio.'
+            : 'Something went wrong connecting to your microphone or camera. Please grant access to those devices to enable audio and video.';
+        // @ts-ignore Only an error is required
+        onError(new Error(msg));
       })
       .finally(() => setIsAcquiringLocalTracks(false));
   }, []);
 
-  const localTracks = [audioTrack, videoTrack, dataTrack].filter(track => track !== undefined) as (
+  // Data tracks don't require browser permissions, so set that up independently of the AV tracks.
+  useEffect(() => {
+    setDataTrack(new LocalDataTrack());
+  }, []);
+
+  const localTracks = [audioTrack, videoTrack, dataTrack].filter((track) => track !== undefined) as (
     | LocalAudioTrack
     | LocalVideoTrack
     | LocalDataTrack